requirements.txt
flask==2.3.3
gunicorn==21.2.0
pandas==2.0.3
numpy==1.24.3
tensorflow==2.13.0
requests==2.31.0
pybit==5.11.0
python-dotenv==1.0.0
flask-sqlalchemy==3.0.5
flask-login==0.6.3
flask-cors==4.0.0
